Move bot into thread, fix rounf and pass button functionality, disable clock
