+++ src/client/ClientFXML.fxml
@@ -26,7 +26,7 @@
 <?import javafx.scene.shape.Circle?>
 <?import javafx.scene.text.Font?>
 
-<StackPane minHeight="-Infinity" minWidth="-Infinity" prefHeight="900.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1">
+<StackPane minHeight="-Infinity" minWidth="-Infinity" prefHeight="900.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1">
    <children>
       <StackPane fx:id="playerNickNamePanel" prefHeight="150.0" prefWidth="200.0">
          <children>
@@ -1651,7 +1651,7 @@
                   <Insets left="50.0" top="660.0" />
                </StackPane.margin>
             </Button>
-            <Button fx:id="passButton" mnemonicParsing="false" prefHeight="30.0" prefWidth="150.0" text="Pass" StackPane.alignment="TOP_RIGHT">
+            <Button fx:id="passButton" mnemonicParsing="false" onAction="#passButtonOnClick" prefHeight="30.0" prefWidth="150.0" text="Pass" StackPane.alignment="TOP_RIGHT">
                <StackPane.margin>
                   <Insets right="50.0" top="660.0" />
                </StackPane.margin>
