+++ src/server/lobby/IRulesManager.java
@@ -3,15 +3,16 @@ package server.lobby;
 import client.logic.MoveType;
 import server.player.Player;
 import server.board.Coordinates;
+import server.player.PlayerTemplate;
 
 import java.util.ArrayList;
 
 public interface IRulesManager {
     boolean checkMove(Coordinates currentCoordinates, Coordinates newCoordinates, MoveType moveType);
 
-    Player checkWinner(Player player);
+    PlayerTemplate checkWinner(PlayerTemplate player);
 
-    Player checkLooser(Player winner);
+    PlayerTemplate checkLooser(PlayerTemplate winner);
 
     void setMediator(LobbyMediator mediator);
 }
