+++ src/server/lobby/LobbyMediator.java
@@ -1,35 +1,33 @@
 package server.lobby;
 
 import client.logic.MoveType;
-import server.player.Player;
+import server.board.Board;
 import server.board.Coordinates;
 import server.board.Field;
-import server.board.IBoard;
 import server.player.PlayerTemplate;
 
-import java.util.ArrayList;
-
 public class LobbyMediator {
-    private Clock clock;
-    private Thread clockThread;
+    private volatile Clock clock;
+    private volatile Thread clockThread;
     private IRulesManager rulesManager;
-    private IBoard board;
+    private Board board;
     private Lobby lobby;
 
     public LobbyMediator() {
     }
 
-    public long getRoundTime() {
-        return clock.getRoundTime();
+    public synchronized long getRoundTime() {
+        return 30;
+        //return clock.getRoundTime();
     }
 
-    public void setBoard(IBoard board) {
+    public void setBoard(Board board) {
         this.board = board;
     }
 
-    public void startRound() {
+    public synchronized void startRound() {
         lobby.sendTurnNotification();
-        clock.startRound();
+       // clock.startRound();
     }
 
     public Field getField(Coordinates coordinates) {
@@ -37,19 +35,19 @@ public class LobbyMediator {
         return field;
     }
 
-    public void endRound() {
-        clock.endRound();
+    public synchronized void endRound() {
+        //clock.endRound();
     }
 
-    public boolean checkMove(Coordinates currentCoordinates, Coordinates newCoordinates, MoveType moveType) {
+    public synchronized boolean checkMove(Coordinates currentCoordinates, Coordinates newCoordinates, MoveType moveType) {
         return rulesManager.checkMove(currentCoordinates, newCoordinates, moveType);
     }
 
-    public boolean move(PlayerTemplate player, Coordinates currentCoordinates, Coordinates newCoordinates) {
+    public synchronized boolean move(PlayerTemplate player, Coordinates currentCoordinates, Coordinates newCoordinates) {
         return board.executeMove(player, currentCoordinates, newCoordinates);
     }
 
-    public void endOfTime() {
+    public synchronized void endOfTime() {
         lobby.nextRound();
     }
 
@@ -57,7 +55,7 @@ public class LobbyMediator {
         this.lobby = lobby;
     }
 
-    public void setClock(Clock clock) {
+    public synchronized void setClock(Clock clock) {
         if(this.clock == null) {
             this.clock = clock;
             clock.setMediator(this);
@@ -72,12 +70,12 @@ public class LobbyMediator {
     }
 
     public void checkWinner() {
-        Player winner = rulesManager.checkWinner(lobby.round);
+        PlayerTemplate winner = rulesManager.checkWinner(lobby.round);
         if(winner != null) {
             //TODO - send popup to winner
             int loosingCorner = (winner.corner + 3) % 6;
-            Player looser = null;
-            for(Player p : lobby.players) {
+            PlayerTemplate looser = null;
+            for(PlayerTemplate p : lobby.players) {
                 if(p.corner == loosingCorner) {
                     looser = p;
                     break;
